pyproject.toml
[tool.poetry]
name = "r-place-blender"
version = "0.1.0"
description = "r/place visualized using Python and Blender"
authors = ["Christopher Kumm <chrislkumm@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"
pandas = "^1.4.2"
numpy = "^1.22.3"
matplotlib = "^3.5.1"
jupyter = "^1.0.0"
black = { version = "^22.3.0", optional = true }
pyarrow = "^7.0.0"
progressbar2 = "^4.0.0"

[tool.poetry.dev-dependencies]

[tool.poetry.extras]
black = ["black"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
